pyproject.toml
[tool.poetry]
name = "project-chiral-ai-service"
version = "0.1.0"
description = ""
authors = ["王博伟 <wangbowei.114514@bytedance.com>"]
readme = "README.md"
packages = [{ include = "project_chiral_ai_service" }]

[tool.poetry.dependencies]
python = "^3.8"
farm-haystack = { extras = ["faiss"], version = "^1.15.0" }
pytest = "^7.2.2"
numpy = "1.23.0"
python-dotenv = "^1.0.0"
redis = "^4.5.1"
pika = "^1.3.1"
pydantic = "^1.10.6"
spacy = "^3.5.1"
en-core-web-trf = { path = "project_chiral_ai_service/models/en_core_web_trf-3.5.0-py3-none-any.whl" }
zh-core-web-trf = { path = "project_chiral_ai_service/models/zh_core_web_trf-3.5.0-py3-none-any.whl" }

[[tool.poetry.source]]
name = "mirrors"
url = "http://mirrors.aliyun.com/pypi/simple/"
default = true
secondary = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
